Separated cleaning data from table logic
